refactor cacher
